Add implementation for 'search' command
This code is probably really messy, this is part of the development for more methods to access the MPR, but I'm also using it as part of my method for learning Rust. So no guarantees on if something isn't looking right :).

We still need to sort searches alphabetically via package names, but we'll do that later.